+++ sw/airborne/subsystems/navigation/common_nav.h
@@ -48,9 +48,9 @@ struct point {
 #define WaypointAlt(_wp) (waypoints[_wp].a)
 #define Height(_h) (_h + ground_alt)
 
-extern void nav_move_waypoint(uint8_t wp_id, float utm_east, float utm_north, float alt);
+extern void nav_move_waypoint(uint16_t wp_id, float utm_east, float utm_north, float alt);
 
-extern const uint8_t nb_waypoint;
+extern const uint16_t nb_waypoint;
 extern struct point waypoints[];
 /** size == nb_waypoint, waypoint 0 is a dummy waypoint */
